datascience-notebook
This commit cleans a few things up to work with the latest Jupyterhub and
also leans on the jupyter/datascience-notebook Docker container which
dockerspawner spins up for logged in users. I removed the python role since
Python is now installed with the system role, and scipy and things like that
come with the datascience-notebook.